Rename utils files to hex
