+++ test/unittests/end_to_end_test.cpp
@@ -1,6 +1,6 @@
 #include "execute.hpp"
+#include "hex.hpp"
 #include "parser.hpp"
-#include "utils.hpp"
 #include <gtest/gtest.h>
 
 using namespace fizzy;
